requirements.txt
mujoco>=3.1.0
numpy>=1.26
PyYAML>=6.0
matplotlib>=3.8
